+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # HidingBar
 
 * Hiding when clicking on a free space.
+* Added "Direction of buttons" option.